@@ -2,9 +2,10 @@ package com.clothing.management.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
+@Table(name = "offer_model")
 public class OfferModel {
 
     @EmbeddedId
