@@ -2,11 +2,12 @@ package com.clothing.management.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.Date;
 import java.util.Set;
 
 @Entity
+@Table(name="packet_status")
 public class PacketStatus {
 
     @Id
