@@ -1,4 +1,4 @@
-package com.clothing.management.config;
+package com.clothing.management.auth.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
