@@ -1,31 +1,32 @@
 package com.clothing.management.repository.repositoryImpl;
 
 import com.clothing.management.entities.Packet;
-import com.clothing.management.enums.DiggieStatus;
+import com.clothing.management.repository.enums.DiggieStatus;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import jakarta.persistence.TypedQuery;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.persistence.criteria.CriteriaQuery;
+import jakarta.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Root;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.TypedQuery;
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-@Component
+@Repository
 public class PacketRepositoryImpl {
 
     @PersistenceContext
     private EntityManager entityManager;
 
-    public Page<Packet> findAllPackets(String searchText, String startDate, String endDate, String status, Pageable pageable) {CriteriaBuilder criteriaBuilder = entityManager.getCriteriaBuilder();
+    public Page<Packet> findAllPackets(String searchText, String startDate, String endDate, String status, Pageable pageable) {
+        CriteriaBuilder criteriaBuilder = entityManager.getCriteriaBuilder();
         CriteriaQuery<Packet> criteriaQuery = criteriaBuilder.createQuery(Packet.class);
         Root<Packet> root = criteriaQuery.from(Packet.class);
 
