@@ -16,7 +16,8 @@ import java.util.Optional;
 public interface PacketService {
 
     public List<Packet> findAllPackets();
-    public Page<Packet> findAllPackets(String searchText, String startDate, String endDate, String status, Pageable pageable);
+    public Page<Packet> findAllPackets(Pageable pageable, String searchText, String startDate, String endDate, String status) throws ParseException;
+    //public Page<Packet> findAllPackets(String searchText, String startDate, String endDate, String status, Pageable pageable);
     //public Page<Packet> findAllPackets(String searchText, int page, int size);
     public Page<Packet> findAllTodaysPackets(Pageable paging);
     public List<Packet> findAllPacketsByDate(String start,String end);
