@@ -1,7 +1,8 @@
 export const baseUrl = "http://localhost:2233";
 //export const APPNAME = "DIGGIE";
 //export const baseUrl = "https://diggie.jvmhost.net/cmb";
-export const APPNAME = "Lyft";
+export const LYFT = "Lyft";
+export const DIGGIE = "Diggie";
 //export const baseUrl = "https://lyft.jvmhost.net/lyft";
 //export const baseUrlfirst = "https://diggie.jvmhost.net/apifirst";
 //export let deliveryAPi = "first"
